call save
